[DRUP-600] add functional js test for the add credit button
