+++ modules/add_credit/tests/src/Kernel/Plugin/AddCreditEntityTypeManagerTest.php
@@ -17,7 +17,7 @@
  * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  */
 
-namespace Drupal\Tests\apigee_m10n_add_credit\Kernel;
+namespace Drupal\Tests\apigee_m10n_add_credit\Kernel\Plugin;
 
 use Drupal\apigee_m10n_add_credit\Plugin\AddCreditEntityTypeInterface;
 use Drupal\Tests\apigee_m10n\Kernel\MonetizationKernelTestBase;
